pyproject.toml
[project]
name = "pandas-pipeline-agent-flow-generator"
version = "0.1.0"
description = "A Langgraph agent flow generator for Pandas data pipelines."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "langchain>=0.2.0",
    "langgraph>=0.0.60",
    "groq>=0.9.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
    "tavily-python>=0.3.0",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
    "pandera>=0.25.0",
    # Adicione outras que você pode precisar
    "langchain-groq>=0.3.6",
    "sentence-transformers>=5.0.0",
    "faiss-cpu>=1.11.0.post1",
    "numpy>=2.3.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
# Esta linha diz ao setuptools para procurar pacotes Python dentro do diretório 'src'.
# Ele vai encontrar 'agents', 'core', 'pipelines' como subpacotes.
find = { where = ["src"] }
